Bug fix, telemetry re-vamp, telemetry values in custom switches

+++ src/myeeprom.h
@@ -214,7 +214,7 @@ typedef struct t_TimerMode
 } __attribute__((packed)) TimerMode ;
 
 typedef struct t_PhaseData {
-  int8_t trim[4];     // -500..500 => trim value, 501 => use trim of phase 0, 502, 503, 504 => use trim of phases 1|2|3|4 instead
+  int16_t trim[4];     // -500..500 => trim value, 501 => use trim of phase 0, 502, 503, 504 => use trim of phases 1|2|3|4 instead
   int8_t swtch;       // swtch of phase[0] is not used
   char name[6];
   uint8_t fadeIn:4;
